refactor(test): split app test using csv and bpmn inputs
This should clarify the purpose of each test class.
In addition
  - avoid duplication when checking the output bpmn file
  - introduce assertion to check output svg file
