+++ java/src/test/java/io/process/analytics/tools/bpmn/generator/AppTest.java
@@ -27,7 +27,7 @@ import org.junit.jupiter.api.Test;
 public class AppTest {
 
     @Test
-    public void main_fail_on_unexisting_input_file() throws Exception {
+    public void main_fail_on_non_existing_input_file() throws Exception {
         int returnCode = runApp("input");
 
         assertThat(returnCode).isEqualTo(2);
@@ -35,92 +35,29 @@ public class AppTest {
 
     @Test
     public void main_fail_on_wrong_export_type() throws Exception {
-
         int returnCode = runApp("input", "--output-type", "unknown_export_type");
 
         assertThat(returnCode).isEqualTo(2);
-    }
-
-    @Test
-    public void main_generates_bpmn_output_file_for_A_2_0() throws Exception {
-        checkBPMNGeneration("bpmn/A.2.0.bpmn.xml", "A.2.0_with_diagram.bpmn.xml");
-    }
-
-    private void checkBPMNGeneration(String inputFileName, String outputFileName) throws IOException {
-        String outputPath = outputPath(outputFileName);
-        runApp(input(inputFileName), "-o", outputPath);
-
-        File bpmnFile = new File(outputPath);
-        assertThat(bpmnFile).exists().isFile();
-        assertThat(fileContent(bpmnFile)).contains("<semantic:definitions").contains("xmlns:semantic=\"http://www.omg.org/spec/BPMN/20100524/MODEL\"");
-    }
-
-    @Test
-    public void main_generates_bpmn_output_file_for_A_2_1() throws Exception {
-        checkBPMNGeneration("bpmn/A.2.1.bpmn.xml", "A.2.1_with_diagram.bpmn.xml");
-    }
-
-    @Test
-    public void main_generates_svg_output_file() throws Exception {
-        String outputPath = outputPath("A.2.0_with_diagram.bpmn.svg");
-        runApp(input("bpmn/A.2.0.bpmn.xml"), "--output-type=SVG", "-o", outputPath);
-
-        File svgFile = new File(outputPath);
-        assertThat(svgFile).exists().isFile();
-        assertThat(fileContent(svgFile)).contains("<svg xmlns=\"http://www.w3.org/2000/svg\"");
-    }
-
-    @Test
-    public void main_generates_ascii_output_file() throws Exception {
-        String outputPath = outputPath("02-startEvent_task_endEvent-without-collaboration.bpmn.txt");
-        runApp(input("bpmn/02-startEvent_task_endEvent-without-collaboration.bpmn.xml"), "--output-type=ASCII", "-o", outputPath);
-
-        File asciiFile = new File(outputPath);
-        assertThat(asciiFile).exists().isFile();
-        assertThat(fileContent(asciiFile)).contains("+---");
-    }
-
-    @Test
-    public void main_generates_bpmn_with_simple_discovery_data() throws Exception {
-        String outputPath = outputPath("from_simple_csv_diagram.bpmn.xml");
-        runApp("--input-type=CSV", "--output", outputPath, input("csv/PatientsProcess/nodeSimple.csv"), input("csv/PatientsProcess/edgeSimple.csv"));
-
-        assertOutFile(outputPath);
-    }
-
-    @Test
-    public void main_generates_bpmn_with_gateways() throws Exception {
-        String outputPath = outputPath("from_csv_with_gateways.bpmn");
-        runApp("--input-type=CSV", "--output", outputPath, input("csv/PatientsProcess/gateway_node_simple.csv"), input("csv/PatientsProcess/gateway_edge_simple.csv"));
-
-        assertOutFile(outputPath);
-    }
-
-    @Test
-    public void main_generates_bpmn_with_patients_process_discovery_data() throws Exception {
-        String outputPath = outputPath("from_patients_csv_diagram.bpmn.xml");
-        runApp("--input-type=CSV", "--output", outputPath, input("csv/PatientsProcess/node.csv"), input("csv/PatientsProcess/edge.csv"));
-
-        assertOutFile(outputPath);
     }
 
     // =================================================================================================================
     // UTILS
     // =================================================================================================================
 
-    private static void assertOutFile(String outputPath) throws IOException {
+    public static void assertBpmnOutFile(String outputPath) throws IOException {
         File bpmnFile = new File(outputPath);
         assertThat(bpmnFile).exists().isFile();
         assertThat(fileContent(bpmnFile)).contains("<semantic:definitions").contains("xmlns:semantic=\"http://www.omg.org/spec/BPMN/20100524/MODEL\"");
     }
 
-
-    private static String input(String fileName) {
-        return "src/test/resources/" + fileName;
+    public static void assertSvgOutFile(String outputPath) throws IOException {
+        File svgFile = new File(outputPath);
+        assertThat(svgFile).exists().isFile();
+        assertThat(fileContent(svgFile)).contains("<svg xmlns=\"http://www.w3.org/2000/svg\"");
     }
 
-    private static String outputPath(String fileName) {
-        return "target/test/output/AppTest/" + fileName;
+    public static String input(String fileName) {
+        return "src/test/resources/" + fileName;
     }
 
 }
